fix(solana): repoint diff() import after the SolanaTransactionDiff move
IntentExecutor and StatefulSwapExecutor call SolanaTransaction.diff() to
log what differs when the server rejects a transaction for an invalid
signature - a real call site, not the dead code the extraction assumed.
Update both imports to the function's new package and correct the doc
comment that said otherwise.
